split files in their respective pkgs
