+++ encoding/go.mod
@@ -1,0 +1,8 @@
+module github.com/hashicorp/go-cty-funcs/encoding
+
+go 1.13
+
+require (
+	github.com/zclconf/go-cty v1.2.1
+	golang.org/x/text v0.3.2 // indirect
+)
