+++ uuid/go.mod
@@ -1,0 +1,9 @@
+module github.com/hashicorp/go-cty-funcs/uuid
+
+go 1.13
+
+require (
+	github.com/google/uuid v1.1.1
+	github.com/hashicorp/go-uuid v1.0.2
+	github.com/zclconf/go-cty v1.2.1
+)
